fix: drop ci schema blank string

+++ macros/drop_ci_schema.sql
@@ -4,7 +4,7 @@
 
 {%- macro default__drop_ci_schema() -%}
 {# defaults postgres syntax #}
-    {% set schema_to_drop = env_var('DBT_CI_SCHEMA', 'ci') %}
+    {% set schema_to_drop = env_var('DBT_CI_SCHEMA', 'ci') if env_var('DBT_CI_SCHEMA') == '' %}
     {% set sql %}
         DROP SCHEMA IF EXISTS "{{ schema_to_drop }}" CASCADE;
     {% endset %}
@@ -15,12 +15,18 @@
 {%- endmacro -%}
 
 {%- macro bigquery__drop_ci_schema() -%}
+    {# Retrieve the CI schema name from environment variables, defaulting to 'ci' #}
     {% set schema_to_drop = env_var('DBT_CI_SCHEMA', 'ci') %}
-    {% set sql %}
-        DROP SCHEMA IF EXISTS `{{ schema_to_drop }}` CASCADE;
-    {% endset %}
-    
-    {% do log("Dropping schema (BigQuery syntax): " ~ "'" ~ schema_to_drop ~ "'", info=True) %}
-    {% do run_query(sql) %}
-    {% do log("CI schema dropped.", info=True) %}
+
+    {% if schema_to_drop %}
+        {% set sql %}
+            DROP SCHEMA IF EXISTS `{{ schema_to_drop }}` CASCADE;
+        {% endset %}
+
+        {% do log("Dropping schema (BigQuery syntax): " ~ "'" ~ schema_to_drop ~ "'", info=True) %}
+        {% do run_query(sql) %}
+        {% do log("CI schema dropped.", info=True) %}
+    {% else %}
+        {% do log("DBT_CI_SCHEMA is not set or is empty. Skipping schema drop.", info=True) %}
+    {% endif %}
 {%- endmacro -%}
